Move testing utils to its own folder
This refactoring is motivated by Numpy's testing support package: https://numpy.org/doc/stable/reference/routines.testing.html
