+++ conda/meta.yaml
@@ -36,6 +36,7 @@ test:
     - orion.algo
     - orion.client
     - orion.storage
+    - orion.testing
   commands:
     - orion --help
