+++ tests/unittests/client/test_client.py
@@ -14,9 +14,9 @@ from orion.core.io.database.ephemeraldb import EphemeralDB
 from orion.core.io.database.pickleddb import PickledDB
 from orion.core.utils.exceptions import BranchingEvent, NoConfigurationError, RaceCondition
 from orion.core.utils.singleton import SingletonNotInstantiatedError, update_singletons
-from orion.core.utils.tests import OrionState
 from orion.storage.base import get_storage
 from orion.storage.legacy import Legacy
+from orion.testing import OrionState
 
 create_experiment = orion.client.create_experiment
 workon = orion.client.workon
