+++ tests/unittests/core/conftest.py
@@ -14,7 +14,7 @@ from orion.core.io.convert import (JSONConverter, YAMLConverter)
 import orion.core.io.experiment_builder as experiment_builder
 from orion.core.io.space_builder import DimensionBuilder
 import orion.core.utils.backward as backward
-from orion.core.utils.tests import default_datetime, MockDatetime
+from orion.testing import default_datetime, MockDatetime
 
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 YAML_SAMPLE = os.path.join(TEST_DIR, 'sample_config.yml')
