+++ tests/unittests/core/worker/test_experiment.py
@@ -13,12 +13,12 @@ import pytest
 import orion.core
 from orion.core.io.space_builder import SpaceBuilder
 import orion.core.utils.backward as backward
-from orion.core.utils.tests import OrionState
 import orion.core.worker.experiment
 from orion.core.worker.experiment import Experiment, ExperimentView
 from orion.core.worker.primary_algo import PrimaryAlgo
 from orion.core.worker.trial import Trial
 from orion.storage.base import get_storage
+from orion.testing import OrionState
 
 
 @pytest.fixture()
